epub2pdf/requirements.txt
ebooklib==0.18
reportlab==4.0.4
Pillow==10.0.0
beautifulsoup4==4.12.2
pyinstaller==6.3.0
lxml==4.9.3